Crate poly_commitment

Re-exports

pub use commitment::PolyComm;

Modules

commitment

This module implements Dlog-based polynomial commitment schema. The following functionality is implemented

error

hash_map_cache

ipa

This module contains the implementation of the polynomial commitment scheme called the Inner Product Argument (IPA) as described in Efficient Zero-Knowledge Arguments for Arithmetic Circuits in the Discrete Log Setting

kzg

This module implements the KZG protocol described in the paper Constant-Size Commitments to Polynomials and Their Applications by Kate, Zaverucha and Goldberg, often referred to as the KZG10 paper.

lagrange_basis

pbt_srs

This module defines Property-based tests for the SRS trait. It includes tests regarding methods the SRS trait should implement. It aims to verify the implementation respects the properties described in the documentation of the methods.

precomputed_srs

To prover and verify proofs you need a Structured Reference String (SRS). The generation of this SRS is quite expensive, so we provide a pre-generated SRS in this repo. Specifically, two of them, one for each pasta curve.

utils

Traits

OpenProof

SRS