poly_commitment

Module utils

Structs

ScaledChunkedPolynomial 🔒

A formal sum of the form s_0 * p_0 + ... s_n * p_n where each s_i is a scalar and each p_i is a polynomial. The parameter P is expected to be the coefficients of the polynomial p_i, even though we could treat it as the evaluations.

Enums

DensePolynomialOrEvaluations

Represent a polynomial either with its coefficients or its evaluations

Functions

batch_dlog_accumulator_check

batch_dlog_accumulator_generate

combine_polys

Combine the polynomials using a scalar (polyscale), creating a single unified polynomial to open. This function also accepts polynomials in evaluations form. In this case it applies an IFFT, and, if necessary, applies chunking to it (ie. split it in multiple polynomials of degree less than the SRS size).