Struct poly_commitment::ipa::SRS

pub struct SRS<G> {
    pub g: Vec<G>,
    pub h: G,
    pub lagrange_bases: HashMapCache<usize, Vec<PolyComm<G>>>,
}

Fields
g: Vec<G>

The vector of group elements for committing to polynomials in coefficient form.

h: G

A group element used for blinding commitments

lagrange_bases: HashMapCache<usize, Vec<PolyComm<G>>>

Commitments to Lagrange bases, per domain size

Implementations

impl<G: CommitmentCurve> SRS<G>

Additional methods for the SRS structure

pub fn verify<EFqSponge, RNG>( &self, group_map: &G::Map, batch: &mut [BatchEvaluationProof<'_, G, EFqSponge, OpeningProof<G>>], rng: &mut RNG ) -> boolwhere EFqSponge: FqSponge<G::BaseField, G, G::ScalarField>, RNG: RngCore + CryptoRng, G::BaseField: PrimeField,

This function verifies a batch of polynomial commitment opening proofs. Return true if the verification is successful, false otherwise.

pub unsafe fn create_trusted_setup(x: G::ScalarField, depth: usize) -> Self

This function creates a trusted-setup SRS instance for circuits with number of rows up to depth.

Safety

This function is unsafe because it creates a trusted setup and the toxic waste is passed as a parameter.

impl<G: CommitmentCurve> SRS<G>where <G as CommitmentCurve>::Map: Sync, G::BaseField: PrimeField,

pub fn create_parallel(depth: usize) -> Self

This function creates SRS instance for circuits with number of rows up to depth.

impl<G: CommitmentCurve> SRS<G>

pub fn open<EFqSponge, RNG, D: EvaluationDomain<G::ScalarField>>( &self, group_map: &G::Map, plnms: &'_ [(DensePolynomialOrEvaluations<'_, G::ScalarField, D>, PolyComm<G::ScalarField>)], elm: &[G::ScalarField], polyscale: G::ScalarField, evalscale: G::ScalarField, sponge: EFqSponge, rng: &mut RNG ) -> OpeningProof<G>where EFqSponge: Clone + FqSponge<G::BaseField, G, G::ScalarField>, RNG: RngCore + CryptoRng, G::BaseField: PrimeField, G: EndoCurve,