Trait poly_commitment::SRS

pub trait SRS<G: CommitmentCurve>: Clone {
    // Required methods
    fn max_poly_size(&self) -> usize;
    fn get_lagrange_basis(
        &self,
        domain_size: usize
    ) -> Option<&Vec<PolyComm<G>>>;
    fn blinding_commitment(&self) -> G;
    fn commit(
        &self,
        plnm: &DensePolynomial<G::ScalarField>,
        num_chunks: usize,
        rng: &mut (impl RngCore + CryptoRng)
    ) -> BlindedCommitment<G>;
    fn mask_custom(
        &self,
        com: PolyComm<G>,
        blinders: &PolyComm<G::ScalarField>
    ) -> Result<BlindedCommitment<G>, CommitmentError>;
    fn commit_non_hiding(
        &self,
        plnm: &DensePolynomial<G::ScalarField>,
        num_chunks: usize
    ) -> PolyComm<G>;
    fn commit_evaluations_non_hiding(
        &self,
        domain: D<G::ScalarField>,
        plnm: &Evaluations<G::ScalarField, D<G::ScalarField>>
    ) -> PolyComm<G>;
    fn commit_evaluations(
        &self,
        domain: D<G::ScalarField>,
        plnm: &Evaluations<G::ScalarField, D<G::ScalarField>>,
        rng: &mut (impl RngCore + CryptoRng)
    ) -> BlindedCommitment<G>;
    fn create(depth: usize) -> Self;
    fn add_lagrange_basis(&mut self, domain: D<G::ScalarField>);
    fn size(&self) -> usize;

    // Provided method
    fn mask(
        &self,
        comm: PolyComm<G>,
        rng: &mut (impl RngCore + CryptoRng)
    ) -> BlindedCommitment<G> { ... }
}

Required Methods

fn max_poly_size(&self) -> usize

The maximum polynomial degree that can be committed to

fn get_lagrange_basis(&self, domain_size: usize) -> Option<&Vec<PolyComm<G>>>

Retrieve the precomputed Lagrange basis for the given domain size

fn blinding_commitment(&self) -> G

Get the group element used for blinding commitments

fn commit( &self, plnm: &DensePolynomial<G::ScalarField>, num_chunks: usize, rng: &mut (impl RngCore + CryptoRng) ) -> BlindedCommitment<G>

Commits a polynomial, potentially splitting the result in multiple commitments.

fn mask_custom( &self, com: PolyComm<G>, blinders: &PolyComm<G::ScalarField> ) -> Result<BlindedCommitment<G>, CommitmentError>

Same as SRS::mask except that you can pass the blinders manually.

fn commit_non_hiding( &self, plnm: &DensePolynomial<G::ScalarField>, num_chunks: usize ) -> PolyComm<G>

This function commits a polynomial using the SRS’ basis of size n.

plnm: polynomial to commit to with max size of sections

num_chunks: the number of commitments to be included in the output polynomial commitment The function returns an unbounded commitment vector (which splits the commitment into several commitments of size at most n).

fn commit_evaluations_non_hiding( &self, domain: D<G::ScalarField>, plnm: &Evaluations<G::ScalarField, D<G::ScalarField>> ) -> PolyComm<G>

fn commit_evaluations( &self, domain: D<G::ScalarField>, plnm: &Evaluations<G::ScalarField, D<G::ScalarField>>, rng: &mut (impl RngCore + CryptoRng) ) -> BlindedCommitment<G>

fn create(depth: usize) -> Self

for now needed by snarky-rs

fn add_lagrange_basis(&mut self, domain: D<G::ScalarField>)

fn size(&self) -> usize

Provided Methods

fn mask( &self, comm: PolyComm<G>, rng: &mut (impl RngCore + CryptoRng) ) -> BlindedCommitment<G>

Turns a non-hiding polynomial commitment into a hidding polynomial commitment. Transforms each given <a, G> into (<a, G> + wH, w) with a random w per commitment.

Object Safety

This trait is not object safe.