poly_commitment

Module utils

Enums

DensePolynomialOrEvaluations

Represent a polynomial either with its coefficients or its evaluations

Functions

batch_dlog_accumulator_check

batch_dlog_accumulator_generate

combine_polys

Combine the polynomials using a scalar (polyscale), creating a single unified polynomial to open. This function also accepts polynomials in evaluations form. In this case it applies an IFFT, and, if necessary, applies chunking to it (ie. split it in multiple polynomials of degree less than the SRS size).